basic PID functioning done, but needs sensor filtering and work around zero point

+++ Project/Project Files/Core/Inc/global.h
@@ -43,9 +43,9 @@ struct controlLoopParam{
 	float angleLastError;
 	float angleErrorDer;
 
-	uint32_t kp;
-	uint32_t ki;
-	uint32_t kd;
+	float kp;
+	float ki;
+	float kd;
 
 	int32_t maxOut;
 	int32_t minOut;
